refactoring corona page
