+++ lib/app/modules/states_map/utils/constants.dart
@@ -1,0 +1,4 @@
+import 'package:latlong/latlong.dart';
+
+final kDefaultLatLong = LatLng(-13.516151006814436, -54.849889911711216);
+final kDefaultZoom = 3.789821910858154;
